Reversed moving and renaming "Utils.kt" to solve merge errors.
