+++ app/src/main/java/com/jerboa/ui/components/comment/CommentNodes.kt
@@ -6,12 +6,12 @@ import androidx.compose.foundation.lazy.LazyListState
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.mutableStateListOf
 import androidx.compose.runtime.remember
+import com.jerboa.CommentNodeData
 import com.jerboa.datatypes.CommentView
 import com.jerboa.datatypes.CommunityModeratorView
 import com.jerboa.datatypes.CommunitySafe
 import com.jerboa.datatypes.PersonSafe
 import com.jerboa.db.Account
-import com.jerboa.util.CommentNodeData
 
 @Composable
 fun CommentNodes(
