+++ app/src/main/java/com/jerboa/ui/components/common/AccountHelpers.kt
@@ -3,10 +3,10 @@ package com.jerboa.ui.components.common
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.livedata.observeAsState
+import com.jerboa.PostViewMode
 import com.jerboa.db.Account
 import com.jerboa.db.AccountViewModel
 import com.jerboa.db.AppSettingsViewModel
-import com.jerboa.util.PostViewMode
 
 @Composable
 fun getCurrentAccount(accountViewModel: AccountViewModel): Account? {
