+++ app/src/main/java/com/jerboa/ui/components/common/ImageViewerDialog.kt
@@ -43,7 +43,7 @@ import coil.ImageLoader
 import coil.compose.rememberAsyncImagePainter
 import coil.decode.GifDecoder
 import coil.decode.ImageDecoderDecoder
-import com.jerboa.util.saveBitmap
+import com.jerboa.saveBitmap
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.withContext
