+++ app/src/main/java/com/jerboa/ui/components/common/VoteHelpers.kt
@@ -7,9 +7,9 @@ import androidx.compose.ui.graphics.vector.ImageVector
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.res.vectorResource
 import com.jerboa.R
+import com.jerboa.VoteType
 import com.jerboa.db.Account
 import com.jerboa.ui.theme.muted
-import com.jerboa.util.VoteType
 
 @Composable
 fun <T> VoteGeneric(
