+++ app/src/main/java/com/jerboa/ui/components/community/CommunityViewModel.kt
@@ -9,6 +9,7 @@ import androidx.compose.runtime.setValue
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import arrow.core.Either
+import com.jerboa.VoteType
 import com.jerboa.api.API
 import com.jerboa.api.followCommunityWrapper
 import com.jerboa.api.retrofitErrorHandler
@@ -20,14 +21,13 @@ import com.jerboa.datatypes.PostView
 import com.jerboa.datatypes.SortType
 import com.jerboa.datatypes.api.GetCommunity
 import com.jerboa.db.Account
+import com.jerboa.loginFirstToast
+import com.jerboa.serializeToMap
 import com.jerboa.ui.components.person.blockPersonRoutine
 import com.jerboa.ui.components.post.deletePostRoutine
 import com.jerboa.ui.components.post.fetchPostsRoutine
 import com.jerboa.ui.components.post.likePostRoutine
 import com.jerboa.ui.components.post.savePostRoutine
-import com.jerboa.util.VoteType
-import com.jerboa.util.loginFirstToast
-import com.jerboa.util.serializeToMap
 import kotlinx.coroutines.launch
 
 class CommunityViewModel : ViewModel() {
