+++ app/src/main/java/com/jerboa/ui/components/community/list/CommunityListActivity.kt
@@ -20,13 +20,13 @@ import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.LocalContext
 import androidx.navigation.NavController
+import com.jerboa.DEBOUNCE_DELAY
 import com.jerboa.db.AccountViewModel
 import com.jerboa.db.AppSettingsViewModel
+import com.jerboa.loginFirstToast
 import com.jerboa.ui.components.common.BottomAppBarAll
 import com.jerboa.ui.components.common.getCurrentAccount
 import com.jerboa.ui.components.home.HomeViewModel
-import com.jerboa.util.DEBOUNCE_DELAY
-import com.jerboa.util.loginFirstToast
 import kotlinx.coroutines.Job
 import kotlinx.coroutines.delay
 import kotlinx.coroutines.launch
