+++ app/src/main/java/com/jerboa/ui/components/home/HomeActivity.kt
@@ -34,19 +34,19 @@ import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.res.stringResource
 import androidx.navigation.NavController
 import com.jerboa.R
+import com.jerboa.VoteType
+import com.jerboa.closeDrawer
 import com.jerboa.db.Account
 import com.jerboa.db.AccountViewModel
 import com.jerboa.db.AppSettingsViewModel
+import com.jerboa.fetchInitialData
+import com.jerboa.loginFirstToast
+import com.jerboa.scrollToTop
 import com.jerboa.ui.components.common.BottomAppBarAll
 import com.jerboa.ui.components.common.getCurrentAccount
 import com.jerboa.ui.components.common.getPostViewMode
 import com.jerboa.ui.components.post.PostListings
 import com.jerboa.ui.components.post.edit.PostEditViewModel
-import com.jerboa.util.VoteType
-import com.jerboa.util.closeDrawer
-import com.jerboa.util.fetchInitialData
-import com.jerboa.util.loginFirstToast
-import com.jerboa.util.scrollToTop
 import kotlinx.coroutines.CoroutineScope
 
 @OptIn(ExperimentalMaterial3Api::class)
