+++ app/src/main/java/com/jerboa/ui/components/person/PersonProfileViewModel.kt
@@ -9,6 +9,7 @@ import androidx.compose.runtime.setValue
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import arrow.core.Either
+import com.jerboa.VoteType
 import com.jerboa.api.API
 import com.jerboa.api.retrofitErrorHandler
 import com.jerboa.datatypes.CommentView
@@ -19,6 +20,8 @@ import com.jerboa.datatypes.SortType
 import com.jerboa.datatypes.api.GetPersonDetails
 import com.jerboa.datatypes.api.GetPersonDetailsResponse
 import com.jerboa.db.Account
+import com.jerboa.serializeToMap
+import com.jerboa.toastException
 import com.jerboa.ui.components.comment.deleteCommentRoutine
 import com.jerboa.ui.components.comment.likeCommentRoutine
 import com.jerboa.ui.components.comment.saveCommentRoutine
@@ -26,9 +29,6 @@ import com.jerboa.ui.components.community.blockCommunityRoutine
 import com.jerboa.ui.components.post.deletePostRoutine
 import com.jerboa.ui.components.post.likePostRoutine
 import com.jerboa.ui.components.post.savePostRoutine
-import com.jerboa.util.VoteType
-import com.jerboa.util.serializeToMap
-import com.jerboa.util.toastException
 import kotlinx.coroutines.launch
 
 class PersonProfileViewModel : ViewModel() {
