+++ app/src/main/java/com/jerboa/ui/components/post/PostRoutines.kt
@@ -3,6 +3,7 @@ package com.jerboa.ui.components.post
 import android.content.Context
 import androidx.compose.runtime.MutableState
 import arrow.core.Either
+import com.jerboa.VoteType
 import com.jerboa.api.deletePostWrapper
 import com.jerboa.api.fetchPostsWrapper
 import com.jerboa.api.likePostWrapper
@@ -12,7 +13,6 @@ import com.jerboa.datatypes.PostView
 import com.jerboa.datatypes.SortType
 import com.jerboa.datatypes.api.DeletePost
 import com.jerboa.db.Account
-import com.jerboa.util.VoteType
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.launch
